17_20260225_OM: Corrected the payment methods amount - ready for release 2.0.0 (#14)
* renamed CheckoutViewModel to GPayCheckoutViewModel

* fixed the payments methods amount to reflect correctly the ui price

* updated the versionName to 2.0.0

+++ app/src/main/java/com/oliviermarteaux/a049_joiefull/Joiefull.kt
@@ -21,7 +21,7 @@ fun Joiefull(
     navController: NavHostController = rememberNavController(),
     windowSize: WindowWidthSizeClass,
     payUiState: PaymentUiState = PaymentUiState.NotStarted,
-    onGooglePayButtonClick: () -> Unit,
+    onGooglePayButtonClick: (Double) -> Unit,
 ) {
 
     val contentType: SharedContentType = when (windowSize){
